create alertDialog and Intent for RecipeDetailActivity

+++ app/src/main/java/com/ilyakoz/saladbowl/cleancode/application/presentation/listRecipe/MainActivity.kt
@@ -3,9 +3,11 @@ package com.ilyakoz.saladbowl.cleancode.application.presentation.listRecipe
 import android.os.Bundle
 import android.view.LayoutInflater
 import androidx.activity.viewModels
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.viewModelScope
 import androidx.recyclerview.widget.GridLayoutManager
+import com.ilyakoz.saladbowl.RecipeDetailActivity
 import com.ilyakoz.saladbowl.cleancode.application.presentation.adapter.SaladBowlAdapter
 import com.ilyakoz.saladbowl.cleancode.application.presentation.createNewRecipe.CreateRecipeActivity
 import com.ilyakoz.saladbowl.databinding.ActivityMainBinding
@@ -57,15 +59,34 @@ class MainActivity : AppCompatActivity() {
 
     private fun deleteRecipeItem() {
         adapter.onRecipeItemLongClickListener = { recipeItem ->
-            viewModel.viewModelScope.launch {
-                viewModel.deleteRecipeItem(recipeItem)
+            val alertDialogBuilder = AlertDialog.Builder(this)
+            alertDialogBuilder.setTitle("Delete Recipe")
+            alertDialogBuilder.setMessage("Are you sure you want to delete this recipe?")
+            alertDialogBuilder.setPositiveButton("Yes") { dialog,_ ->
+                viewModel.viewModelScope.launch {
+                    viewModel.deleteRecipeItem(recipeItem)
+                }
+                dialog.dismiss()
             }
+            alertDialogBuilder.setNegativeButton("No") {dialog,_ ->
+                dialog.dismiss()
+            }
+            val alertDialog = alertDialogBuilder.create()
+            alertDialog.show()
+
         }
     }
 
-    private fun getRecipeItem() {
+//    private fun getRecipeItem() {
+//        adapter.onRecipeItemClickListener = {
+//            val intent = CreateRecipeActivity.newIntentEditItem(this, it.id)
+//            startActivity(intent)
+//        }
+//    }
+
+    private fun getRecipeItem(){
         adapter.onRecipeItemClickListener = {
-            val intent = CreateRecipeActivity.newIntentEditItem(this, it.id)
+            val intent = RecipeDetailActivity.newIntentShowDetailRecipe(this, it.id)
             startActivity(intent)
         }
     }
